Extract Renderer from explain Summary
Condition/expression stringification (precedence, clarity parens, field
naming) moves to its own class.
